Fix: change the type field for the foreign key news_id on news_comments table

+++ database/migrations/2022_07_29_064356_create_news_comments_table.php
@@ -17,7 +17,7 @@ class CreateNewsCommentsTable extends Migration
             $table->increments('id');
             $table->timestamps();
             $table->text('comment_text')->collation('utf8mb4_bin');
-            $table->unsignedBigInteger('new_id');
+            $table->unsignedInteger('new_id');
             $table->unsignedInteger('user_id');
             });
     }
